move curve fitting+tests, edit mobility function names
